requirements.txt
asgiref==3.11.1
Django==6.0.2
pillow==12.1.0
sqlparse==0.5.5
tzdata==2025.3
dj-database-url==2.1.0
python-dotenv==1.0.1
whitenoise==6.4.0
django-storages[boto3]==2.2.0
boto3==1.30.24
psycopg[binary]
gunicorn